Moved config to config package
